refactor(sync): extract shared stream-bootstrap retry + query options
Self-review pass on the previous commit:
- Move bootstrapRetry/bootstrapRetryDelay + retry constants out of
  use-coordinated-stream-queries.ts into a new lib/stream-bootstrap-query.ts
  so use-streams.ts no longer cross-imports from a sibling hook for what is
  really a cross-cutting concern.
- Add STREAM_BOOTSTRAP_QUERY_OPTIONS to deduplicate the 8 identical query
  options (staleTime/gcTime/retry/retryDelay/refetchOn*/structuralSharing)
  shared between useCoordinatedStreamQueries and useStreamBootstrap, so the
  two observers on the same query key cannot drift.
- Tighten redundant comments on the retry/enabled lines and drop a
  self-explanatory test setup comment.
- Move bootstrapRetry/bootstrapRetryDelay unit tests next to the helpers.
