+++ apps/frontend/src/hooks/use-streams.ts
@@ -2,7 +2,7 @@ import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query"
 import { useSocket, useStreamService } from "@/contexts"
 import { debugBootstrap } from "@/lib/bootstrap-debug"
 import { getQueryLoadState, isTerminalBootstrapError } from "@/lib/query-load-state"
-import { bootstrapRetry, bootstrapRetryDelay } from "./use-coordinated-stream-queries"
+import { STREAM_BOOTSTRAP_QUERY_OPTIONS } from "@/lib/stream-bootstrap-query"
 import { db } from "@/db"
 import { joinRoomBestEffort } from "@/lib/socket-room"
 import { applyStreamBootstrap, toCachedStreamBootstrap, type CachedStreamBootstrap } from "@/sync/stream-sync"
@@ -103,20 +103,10 @@ export function useStreamBootstrap(workspaceId: string, streamId: string, option
         { incrementWindowVersionOnReplace: bootstrap.syncMode === "replace" }
       )
     },
-    // Keep terminal auth/not-found errors disabled to avoid loops.
-    // Non-terminal errors can recover automatically via retry below.
+    // Terminal 403/404 disables the query to prevent loops; recoverable errors
+    // self-heal via STREAM_BOOTSTRAP_QUERY_OPTIONS.retry.
     enabled: (options?.enabled ?? true) && !!workspaceId && !!streamId && !!socket && !hasTerminalError,
-    // Match coordinated loading options to share cache correctly and prevent
-    // multiple observers from conflicting. Coordinated loading handles initial
-    // fetch, and socket events handle updates.
-    staleTime: Infinity,
-    gcTime: Infinity,
-    retry: bootstrapRetry,
-    retryDelay: bootstrapRetryDelay,
-    refetchOnMount: false,
-    refetchOnWindowFocus: false,
-    refetchOnReconnect: false,
-    structuralSharing: false,
+    ...STREAM_BOOTSTRAP_QUERY_OPTIONS,
   })
 
   const loadState = getQueryLoadState(query.status, query.fetchStatus)
